Rename `TextualEvent` to `TextualEventView`
